Update Quick Start Guide
The Quick Start Guide previously included a required Sample field in the description of Step 2 (Upload Assay). However, Sample records are no longer associated with the Assay record and are not designated during the Upload Assay workflow. The Text of the Quick Start Guide was modified to reflect this, and the screenshot of Upload Assay was updated to the newest version as well.