Resolve data loading issue
Data was not loaded correctly into the well edit form. This changes the javascript code to use a novel variable name for the new field, and changes the form html such that the 'name' and 'id' attributes of the new field are equal and referenced correctly by the javascript code. Furthermore, the correct key for sample concentration is now referenced (was 'contents' but should be 'sample_conc'). This was causing the javascript file to silently fail and not load data into the rest of the fields.

This commit appears to resolve data loading issues into the Well Edit form, but does not yet resolve data saving issues which still persist.