Resolve Well Edit data saving issues
Well edit was not correctly saving data. Nothing appeared to save correctly, although troubleshooting was complicated by two issues: (1) It may take a long time for the SQL to time out and for an error to be displayed, and (2) Sometimes the browser cache makes it look like changes have been pushed to the database when they are not. This fixes a line in the SQL query Q_LOAD_WELL_UPDATES, which specifies the columm names in the temporary CSV file used for data upload. The file was correctly changed, but the specification of column names was not updated in the "ENCLOSED BY..." section of this query.