Added  argument "maxval" to drawChart function. Call from drawChartGrid() with arbitrary value "200,000" to observe change in graphs. This "200,000" should eventually be replaced with the maxvalue.
Added vAxis viewWindow min and max values in options{}.

Defined variable maxval when creating an assay. Loop through numbers in fluorescence_series and update maxval when a number is larger.

QUESTIONS:
where should I store this value so that it can be accessed when creating graphs AND so that it remains unique for each created assay?

Should this field be added to Assay in the db?
 What are the javascript queries accessing when they pull Assay_ID etc? Should this value go there?